+++ R/utils_get_representative_values.R
@@ -118,7 +118,7 @@
 
   # check if levels were numeric or not...
   suppressWarnings(
-    if (!anyNA(unlist(lapply(at_levels, as.numeric)))) {
+    if (!anyNA(unlist(lapply(at_levels, as.numeric))) && !any(lapply(at_levels, .is_pseudo_numeric))) {
       at_levels <- lapply(at_levels, as.numeric)
     }
   )